Adjusted how the note hit window works.
Before there were roughly 66ms for the early window and roughly 134ms for the late window (200ms total). Now with new input, the early and late windows are 100ms each. This makes it so you can actually hit BAD and SHIT note early as well as making the note timing more consistent in general. I might adjust the hit window a bit based around this change.