Input stuff.
I think I'm gonna stick with the method I came up  with and leave listener stuff commented out in the code. Gonna have some people test this and if they're input is fine and they don't have drops, I think I'll keep it since this works with keyboard and controller.