Discount Paths.hx and better chart editor Cval proofing.
Adds a Paths.hx that works similarly to the normal one from the new assets structure version of the game. I will check if an image is cached and only load from file if it isn't. Also adds a more elegant solution to Cval's double note issue and will automatically fix charts if they are opened in the editor.