welcome to a new era mister bud
new health icon system to make it easier to work with + it doesn't crash if you forget to set an icon. setting an icon is done my the character instead of the health icon. also retires boyfriend.hx and moves all relevant code to either character or gameover. im setting up a decent rewrite of character.hx to make it easier to work with sprites that get flipped + make offsets work in both directions.