BIG UPDATE TO CHARACTER SCALE AND LOOPING
character offsets are now set at the original scale of the character and the offset will be automatically fixed whenever the character is scaled. loop points are now defined by the character instead of being hard coded in the character anim end function. loop points are done manually so the anim end function actually runs