Update Highscore.hx
Instead of creating a blank score and saving it to disk when a score isn't available it just returns the blank score and doesn't write to disk. This makes it so that first time players that open freeplay don't freeze for a few seconds while the game writes like 120 scores to your save individually.