fixed weird music bug
fixes a bug that could cause song playback to fuck up when repeatedly pausing or tabbing out. caused by the game continuously allocating memory when doing those things and it would just not reclaim it sometimes. thank u @GlitchDotSmh on twitter for bringing this to my attention and helping me test the fix for ti