start main menu rewrite
It uses json files to define menu buttons so mods can add custom buttons to the menu. I need to add the ability for it to scroll and also make it look nice with less than 4 buttons but I think it's fully functional in terms of recreating the normal menu stuff. Also I need to make freeplay do this shit also but that should just be a copy past when this is all done. (Also maybe make the default menu stuff built-in so mods can't fuck them up? I don't know about this one yet.)